avoid warning

+++ YumiMediationAdapters/GDT/YumiMediationBannerAdapterGDT.m
@@ -72,7 +72,7 @@
                                             placementId:weakSelf.provider.data.key2 ?: @""
                                          viewController:[weakSelf.delegate rootViewControllerForPresentingModalView]];
         weakSelf.unifiedBannerView.animated = NO;
-        weakSelf.unifiedBannerView.autoSwitchInterval = weakSelf.provider.data.autoRefreshInterval;
+        weakSelf.unifiedBannerView.autoSwitchInterval = (int)weakSelf.provider.data.autoRefreshInterval;
         weakSelf.unifiedBannerView.delegate = weakSelf;
         
         [weakSelf.unifiedBannerView loadAdAndShow];
